fix loss normalization for ddp training

+++ funasr/models/e2e_asr_contextual_paraformer.py
@@ -233,7 +233,7 @@ class NeatContextualParaformer(Paraformer):
         stats["loss"] = torch.clone(loss.detach())
         # force_gatherable: to-device and to-tensor if scalar for DataParallel
         if self.length_normalized_loss:
-            batch_size = (text_lengths + self.predictor_bias).sum().type_as(batch_size)
+            batch_size = int((text_lengths + self.predictor_bias).sum())
         loss, stats, weight = force_gatherable((loss, stats, batch_size), loss.device)
         return loss, stats, weight
